scout/requirements.txt
streamlit>=1.35.0
praw>=7.7.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.30.0
pandas>=2.0.0
watchdog>=4.0.0
